+++ larmoby/src/main/java/com/example/larmoby/repository/ItemPedidoRepository.java
@@ -13,8 +13,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-
-public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Long>  {
+public interface ItemPedidoRepository extends JpaRepository<ItemPedido, Integer> {
 
     @Query("SELECT i FROM ItemPedido i WHERE i.id_item_pedido = ?1")
     List<ItemPedido> findItemPedidoById_pedido(int id_pedido);
